Improve android wear app (#257)
* Added timer and end summary to wear activity
* Added timber
* extracted getUriForFile to utils
* Added data binding to wear app
* Updated translations
* Cleaned up travis file

+++ shared/src/main/java/de/dreier/mytargets/shared/utils/EndRenderer.java
@@ -87,7 +87,7 @@ public class EndRenderer {
         } while (neededRows > maxRows);
         radius -= MIN_PADDING;
         int numRows = Math.max(neededRows, 1);
-        shotsPerRow = (int) Math.ceil(shotList.size() / numRows);
+        shotsPerRow = (int) Math.ceil(shotList.size() / (float)numRows);
         rowHeight = rect.height() / numRows;
         columnWidth = rect.width() / shotsPerRow;
     }
